--- src/Untitled-1
@@ -1,164 +1,0 @@
-yarn add ts-protoc-gen
-yarn add grpc_tools_node_protoc_ts
-
-protoc --plugin=./node_modules/.bin/protoc-gen-ts --js_out=import_style=commonjs,binary:src/proto --ts_out=service=true:src/proto -I ./../iroha/shared_model/schema ../iroha/shared_model/schema/*.proto
-
-protoc --js_out=import_style=commonjs,binary:src/proto  -I ./../iroha/shared_model/schema ../iroha/shared_model/schema/*.proto
-
-grpc_tools_node_protoc --js_out=import_style=commonjs,binary:src/proto --grpc_out=src/proto --plugin=protoc-gen-grpc=`which grpc_tools_node_protoc_plugin` -I ./../iroha/shared_model/schema ../iroha/shared_model/schema/*.proto
-
-block_pb.js
-block_pb_service.js
-
-commands_pb.js
-commands_pb_service.js
-
-endpoint_pb.js
-endpoint_pb_service.js
-
-primitive_pb.js
-primitive_pb_service.js
-
-proposal_pb.js
-proposal_pb_service.js
-
-qry_responses_pb.js
-qry_responses_pb_service.js
-
-queries_pb.js
-queries_pb_service.js
-
-transaction_pb.js
-transaction_pb_service.js
-
-
-
-
-
-###########
-###########
-###########
-// for usage with grpc package use endpoint_grpc_pb file
-import grpc from 'grpc'
-import {
-  QueryService_v1Client,
-  CommandService_v1Client
-} from '../lib/proto/endpoint_grpc_pb'
-import { queryHelper, txHelper } from '../lib'
-
-import { TxStatus, TxStatusRequest } from '../lib/proto/endpoint_pb'
-
-import flow from 'lodash.flow'
-
-let irohaAddress = 'localhost:50051'
-
-let adminPriv =
-  'f101537e319568c765b2cc89698325604991dca57b9716b58016b253506cab70'
-
-// Creating transaction with txHelper
-let transaction = flow(
-  t =>
-    txHelper.addCommand(t, 'CreateAsset', {
-      assetName: 'dollar10',
-      domainId: 'test',
-      precision: 2
-    }),
-  t =>
-    txHelper.addMeta(t, {
-      creatorAccountId: 'admin@test'
-    })
-)(txHelper.emptyTransaction())
-
-let transaction2 = flow(
-  t =>
-    txHelper.addCommand(t, 'CreateAsset', {
-      assetName: 'dollar11',
-      domainId: 'test',
-      precision: 2
-    }),
-  t =>
-    txHelper.addMeta(t, {
-      creatorAccountId: 'admin@test'
-    })
-)(txHelper.emptyTransaction())
-
-// for usage with grpc package don't forget to pass credentials or grpc.credentials.createInsecure()
-const txClient = new CommandService_v1Client(
-  irohaAddress,
-  grpc.credentials.createInsecure()
-)
-
-const batchArray = txHelper.addBatchMeta([transaction, transaction2], 0)
-
-batchArray[0] = txHelper.sign(batchArray[0], adminPriv)
-batchArray[1] = txHelper.sign(batchArray[1], adminPriv)
-
-const txHash = txHelper.hash(batchArray[0])
-const txHash2 = txHelper.hash(batchArray[1])
-
-const batch = txHelper.createTxListFromArray(batchArray)
-
-txClient.listTorii(batch, (err, data) => {
-  if (err) {
-    throw err
-  } else {
-    console.log(
-      'Submitted transaction successfully! Hash: ' +
-        txHash.toString('hex') + '\n' +
-        txHash2.toString('hex')
-    )
-  }
-})
-
-const request = new TxStatusRequest()
-
-request.setTxHash(txHash.toString('hex'))
-
-let stream = txClient.statusStream(request)
-
-stream.on('data', function (response) {
-  console.log(response.getTxStatus())
-})
-
-stream.on('end', function (end) {
-  console.log('finish')
-})
-
-const request2 = new TxStatusRequest()
-
-request2.setTxHash(txHash2.toString('hex'))
-
-let stream2 = txClient.statusStream(request2)
-
-stream2.on('data', function (response) {
-  console.log(response.getTxStatus())
-})
-
-stream2.on('end', function (end) {
-  console.log('finish')
-})
-
-// Creating query with queryHelper
-let query = flow(
-  (q) => queryHelper.addQuery(q, 'getAccount', { accountId: 'admin@test' }),
-  (q) => queryHelper.addMeta(q, { creatorAccountId: 'admin@test' }),
-  (q) => queryHelper.sign(q, adminPriv)
-)(queryHelper.emptyQuery())
-
-// for usage with grpc package don't forget to pass credentials or grpc.credentials.createInsecure()
-const queryClient = new QueryService_v1Client(
-  irohaAddress,
-  grpc.credentials.createInsecure()
-)
-
-// Sending query with queryHelper
-queryClient.find(query, (err, response) => {
-  if (err) {
-    throw err
-  } else {
-    console.log(JSON.stringify(response))
-  }
-})
-#########
-#########
-#########
